+++ MqttTopicBuilder.UnitTests/Builder/TopicBuilderUnitTests.cs
@@ -9,9 +9,12 @@
  *      MIT - https://github.com/pBouillon/MqttTopicBuilder/blob/master/LICENSE
  */
 
+using System;
 using AutoFixture;
 using FluentAssertions;
 using MqttTopicBuilder.Builder;
+using MqttTopicBuilder.Constants;
+using MqttTopicBuilder.Exceptions.Classes;
 using MqttTopicBuilder.UnitTests.Utils;
 using Xunit;
 
@@ -55,6 +58,30 @@ namespace MqttTopicBuilder.UnitTests.Builder
         }
 
         /// <summary>
+        /// Ensure that the object's cleaning behaviour is valid
+        /// </summary>
+        [Fact]
+        public void Clear()
+        {
+            // Arrange
+            var builder = Fixture.Create<TopicBuilder>();
+
+            // Act
+            var cleaned = builder.Clear();
+
+            // Assert
+            cleaned.Levels
+                .Should()
+                .Be(0,
+                    "because the builder must not contain any level anymore");
+
+            cleaned.TopicCollection
+                .ToList()
+                .Should()
+                .BeEmpty("because the inner collection should have been cleaned");
+        }
+
+        /// <summary>
         /// Ensure that the clone is successful
         /// </summary>
         [Fact]
@@ -94,6 +121,52 @@ namespace MqttTopicBuilder.UnitTests.Builder
             clone.Levels.Should()
                 .Be(initialCount,
                     "because altering the origin should not alter the cloned instance");
+        }
+
+        /// <summary>
+        /// Ensure that a <see cref="ITopicBuilder"/> can be created from an existing topic
+        /// </summary>
+        [Fact]
+        public void FromTopic()
+        {
+            // Arrange
+            var topic = MqttTopicBuilder.Builder.Topic.FromString(
+                TestUtils.GenerateValidTopic());
+
+            // Act
+            var builder = TopicBuilder.FromTopic(topic, TopicConsumer.Subscriber);
+
+            // Assert
+            builder.Build()
+                .Should()
+                .Be(topic, 
+                    "because the topic should have been built from the provided one");
+
+            builder.MaxLevel
+                .Should()
+                .Be(topic.Levels,
+                    "because the builder should only hold the provided topic");
+        }
+
+        /// <summary>
+        /// Ensure that it is not possible to bypass construction rule on a specific <see cref="TopicConsumer"/>
+        /// by seeding the <see cref="TopicBuilder"/> with a <see cref="Topic"/> containing forbidden values
+        /// </summary>
+        [Fact]
+        public void FromTopic_WithIllegalConsumer()
+        {
+            // Arrange
+            var wildcardTopic = MqttTopicBuilder.Builder.Topic.FromString(
+                Mqtt.Wildcard.SingleLevel.ToString());
+
+            // Act
+            Action creatingPublisherBuilderWithSubscriberTopic = () =>
+                _ = TopicBuilder.FromTopic(wildcardTopic, TopicConsumer.Publisher);
+
+            // Assert
+            creatingPublisherBuilderWithSubscriberTopic.Should()
+                .Throw<IllegalTopicConstructionException>(
+                    "because a wildcard should not be allowed in PUB mode");
         }
     }
 }