+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2020-08-26
+
+### Added
+
+- Add `ITopicBuilder.Clear` method
+- Add `TopicBuilder.FromTopic` method
+- Add `Topic.ToArray` method
+- Add `ITopicBuilder` creation from `Topic`
+- Add `ITopicBuilder.ToPublisherBuilder` extension method
+- Add `ITopicBuilder.ToSubscriberBuilder` extension method
+
+### Changed
+
+- `TopicBuilder` no longer use `TopicConsumer.Subscriber` by default
+- `TopicBuilder` now performs rule checks when built with a collection
+
+### Removed
+
+- Remove obsolete `TopicBuilder` constructors
+- Remove unused `ITopicCollection.Clear` method
+
 ## [2.1.0] - 2020-06-26
 
 ### Added